refactor(ci-flaky): split CLI commands
